Add casts for safety
Not certain this would always work with bools

+++ include/UnitCheckedType.hpp
@@ -308,7 +308,7 @@ class UnitCheckedTypeFull{
 */
     static constexpr bool hasNoUnits(){return is_equal(L, 0) && is_equal(M, 0) && is_equal(T, 0) && is_equal(K, 0) && is_equal(A, 0) && is_equal(MO, 0) && is_equal(CD, 0);}///<Check if this type has all Unit exponents zero
 
-    static constexpr bool isFundamentalType(){return ((L==1) + (M==1) + (T==1) + (K==1) + (A==1) + (MO==1) + (CD==1))==1;}///<Check if this type is a fundamental unit type (only one exponent is 1, all others are zero)
+    static constexpr bool isFundamentalType(){return ((int)(L==1) + (int)(M==1) + (int)(T==1) + (int)(K==1) + (int)(A==1) + (int)(MO==1) + (int)(CD==1))==1;}///<Check if this type is a fundamental unit type (only one exponent is 1, all others are zero)
     static constexpr size_t whichFundamentalType(){
         //I am sorry, but the compiler parameter order and the usual SI units order do not match and it's too late to fix. These are the order of the strings in unitNames
         return (L==1)? 1 : (M==1)? 0 : (T==1)? 2 : (K==1)? 3 : (A==1)? 4 : (MO==1)? 5 : 6;}///<Get the index of the fundamental unit type (0-6), junk if not fundamental
